+++ iti-react/Library.d.ts
@@ -2,3 +2,5 @@
 // some intellisense errors
 declare module 'input-format'
 declare module 'react-select'
+
+declare type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>